Optimize svt_search_one_dual_neon function
The tot_mse buffer was initialized to zero, and these values were
loaded and used in the first update loop. Move the first iteration
logic outside of the loop to remove redundant loads and additions.

Unroll the i-loop by factor of one to reduce the number of loads and
stores for the tot_mse buffer.