SadTest: add comment on sad16x16_32b about std::array
The array is in response to gcc 11 emitting the following with rtc+minimal

```c
test/SadTest.cc:923:56: error: 'svt_ext_eight_sad_calculation_32x32_64x64_c' accessing 512 bytes in a region of size 32 [-Werror=stringop-overflow=]
  923 |             svt_ext_eight_sad_calculation_32x32_64x64_c(
      |                                                        ^
test/SadTest.cc:923:56: note: referencing argument 1 of type 'uint32_t[8] *'
Source/Lib/Codec/motion_estimation.c:369: note: in a call to function 'svt_ext_eight_sad_calculation_32x32_64x64_c'
  369 | void svt_ext_eight_sad_calculation_32x32_64x64_c(uint32_t  p_sad16x16[16][8],
      |
```

Based on the warning, it seems that GCC interpreted correctly that the
C function took `uint32_t[16][8]` or 512 bytes total, but thought that
`sad16x16_32b` only had `uint32_t[8]*`, so 4*8, and thought it was only
dereferencing that much. This seems to be due to the C/C++ boundary
leading to the compiler's confusion. It can be confirmed that GCC 15.2
does not emit this warning, but versions in between were not tested.

Signed-off-by: Christopher Degawa <ccom@randomderp.com>