Add CPU part detection for Neoverse V2
Add runtime CPU part detection for Neoverse V2 cores, so they can make
use of an optimized version of sad_loop_kernel that is only benefical
for this core.

This commit reverts the changes of
https://gitlab.com/AOMediaCodec/SVT-AV1/-/merge_requests/2493 and
creates an alternative neoverse_v2.c file that uses the optimization
added in this PR. Only the SVE and Neon Dotprod optimizations (used in
the SVE version) have been retained, as Neoverse V2 cores will always
support these extensions, so there is no need for an alternative Armv8.0
Neon version.