Optimize SVE implementation of av1_warp_affine
In case of beta == 0 and alpha == 0 we know filter values before
processing loops so add new logic using Neon USMMLA instructions,
keeping the SVE implementation for the remaining cases.

By permuting the input samples and the 6-tap filter we can use the
Armv8.6 I8MM USMMLA matrix multiply instructions to accelerate
horizontal 6-tap convolutions. The 2x8 by 8x2 matrix multiply
instruction does twice the work of the USDOT dot product
instructions.

In case of 8-tap filter we can replace the USDOT instruction with
USMMLA, to apply a 7-tap filter, and an extra multiplication.

This is a port from libaom:
https://aomedia-review.googlesource.com/c/aom/+/204462

Change-Id: Ib5068557ca4d3e3b71c8bf08741de91e6b478e43