Optimize Neon I8MM implementation of av1_warp_affine
In case of beta == 0 and alpha == 0 we know filter values before
processing loops so add new logic using USMMLA instruction for
horizontal filtering.

By permuting the input samples and the 6-tap filter we can use the
Armv8.6 I8MM USMMLA matrix multiply instructions to accelerate
horizontal 6-tap convolutions. The 2x8 by 8x2 matrix multiply
instruction does twice the work of a USDOT dot product instruction.

In case of 8-tap filter we can replace the USDOT instruction with
USMMLA, to apply a 7-tap filter, and an extra multiplication.

This is a port from libaom:
https://aomedia-review.googlesource.com/c/aom/+/204461

Change-Id: I7b87e1f0029da16e66387533004d75d4e4008568