emscripten: fix some basic errors with emscripten
This doesn't guarentee an actual working EncApp, just that it builds and
links using:

```bash
emcmake cmake -GNinja -B emscripten \
    -DCMAKE_EXPORT_COMPILE_COMMANDS=ON \
    -DCMAKE_BUILD_TYPE=Release \
    -DBUILD_SHARED_LIBS=OFF \
    -DCMAKE_OUTPUT_DIRECTORY=emscripten \
    -DCMAKE_C_FLAGS="-sMEMORY64 -pthread" \
    -DCMAKE_EXE_LINKER_FLAGS="-sMEMORY64 -sALLOW_MEMORY_GROWTH=1 -sMAXIMUM_MEMORY=4GB -pthread -sNODERAWFS=1"
ninja -C emscripten
```

`-sMEMORY64` is to make sure we are using 64-bit pointers.
`-pthread` is needed to force pthread to actually work instead of just
    throwing resource unavailable errors at runtime.
`-sALLOW_MEMORY_GROWTH=1 -sMAXIMUM_MEMORY=4GB` is for malloc and co to
    work properly. Otherwise, we are stuck with the default 2GB.
`-sNODERAWFS=1` is for `node SvtAv1EncApp.js -i file.y4m` to work
    properly.

specifying `-sMEMORY64` on the commandline is important since
emscripten's cmake toolchain file will hardcode 32-bit if it doesn't
find it in the CMAKE_C_FLAGS, and unfortunately, setting $env{CFLAGS}
is not enough as the toolchain file kicks in before concatenation with
the env variables.

Signed-off-by: Christopher Degawa <1480c1@videolan.org>